Don't do build unless a ChangeLog exists.

+++ adium/Utilities/Build/buildClean.sh
@@ -15,7 +15,7 @@ else
 	lastbuild=$today
 fi
 
-rm -R /Users/evands/AdiumNightly/adium
+rm -R adium
 cvs co adium
 
 cd adium
@@ -31,37 +31,44 @@ fi
 #Log everything
 ../cvs2cl.pl --no-times --day-of-week --prune --hide-filenames --file CompleteChanges
 ../cvs2cl.pl --no-times --day-of-week -l "-d'>=$lastbuild'" --prune --hide-filenames --file ChangeLog
-cp ChangeLog ChangeLog_$prettydate
+if ([ -f ChangeLog ] ) ; then
+    cp ChangeLog ChangeLog_$prettydate
 
-#build it
-xcodebuild -target Adium GENERATE_DEBUGGING_SYMBOLS=NO COPY_PHASE_STRIP=YES DEBUGGING_SYMBOLS=NO MACOSX_DEPLOYMENT_TARGET=10.2 ZERO_LINK=NO OPTIMIZATION_CFLAGS=-Os FIX_AND_CONTINUE=NO OTHER_CFLAGS=-DDEPLOYMENT_BUILD
-#Package it
-../buildDMG.pl -buildDir . -compressionLevel 9 -dmgName "Adium_$prettydate" -volName "Adium_$prettydate" build/Adium.app ChangeLog
+    #build it
+    xcodebuild -target Adium GENERATE_DEBUGGING_SYMBOLS=NO \
+        COPY_PHASE_STRIP=YES DEBUGGING_SYMBOLS=NO \
+        MACOSX_DEPLOYMENT_TARGET=10.2 ZERO_LINK=NO OPTIMIZATION_CFLAGS=-Os \
+        FIX_AND_CONTINUE=NO OTHER_CFLAGS=-DDEPLOYMENT_BUILD
 
-if !([ -e ~/AdiumBuilds ]); then
-    mkdir ~/AdiumBuilds
+
+    #Package it
+    ../buildDMG.pl -buildDir . -compressionLevel 9 -dmgName "Adium_$prettydate" -volName "Adium_$prettydate" build/Adium.app ChangeLog
+
+    if !([ -e ~/AdiumBuilds ]); then
+        mkdir ~/AdiumBuilds
+    fi
+
+    cp Adium_$prettydate.dmg ~/AdiumBuilds/Adium_$prettydate.dmg
+
+    #Copy the files, setting them to be group writeable after copying
+    scp Adium_$prettydate.dmg $username@shell.sf.net:/home/groups/a/ad/adium/htdocs/downloads/
+    ssh shell.sf.net chmod 664 /home/groups/a/ad/adium/htdocs/downloads/Adium_$prettydate.dmg
+
+    scp CompleteChanges $username@shell.sf.net:/home/groups/a/ad/adium/htdocs/downloads/
+    ssh shell.sf.net chmod 664 /home/groups/a/ad/adium/htdocs/downloads/CompleteChanges
+
+    scp ChangeLog_$prettydate $username@shell.sf.net:/home/groups/a/ad/adium/htdocs/downloads/ChangeLogs
+    ssh shell.sf.net chmod 664 /home/groups/a/ad/adium/htdocs/downloads/ChangeLogs/ChangeLog_$prettydate
+
+    ssh shell.sf.net ln -fs \
+        /home/groups/a/ad/adium/htdocs/downloads/Adium_$prettydate.dmg \
+        /home/groups/a/ad/adium/htdocs/downloads/Adium2.dmg
+
+    #cleanup
+    #rm CompleteChanges ChangeLog ChangeLog_$prettydate
+    #rm Adium_$prettydate.dmg
+    cd ..
+    rm .lastadiumbuild
+    echo `date +"%Y-%m-%d"` >> .lastadiumbuild
 fi
-
-cp Adium_$prettydate.dmg ~/AdiumBuilds/Adium_$prettydate.dmg
-
-#Copy the files, setting them to be group writeable after copying
-scp Adium_$prettydate.dmg $username@shell.sf.net:/home/groups/a/ad/adium/htdocs/downloads/
-ssh shell.sf.net chmod 664 /home/groups/a/ad/adium/htdocs/downloads/Adium_$prettydate.dmg
-
-scp CompleteChanges $username@shell.sf.net:/home/groups/a/ad/adium/htdocs/downloads/
-ssh shell.sf.net chmod 664 /home/groups/a/ad/adium/htdocs/downloads/CompleteChanges
-
-scp ChangeLog_$prettydate $username@shell.sf.net:/home/groups/a/ad/adium/htdocs/downloads/ChangeLogs
-ssh shell.sf.net chmod 664 /home/groups/a/ad/adium/htdocs/downloads/ChangeLogs/ChangeLog_$prettydate
-
-ssh shell.sf.net ln -fs \
-    /home/groups/a/ad/adium/htdocs/downloads/Adium_$prettydate.dmg \
-    /home/groups/a/ad/adium/htdocs/downloads/Adium2.dmg
-
-#cleanup
-#rm CompleteChanges ChangeLog ChangeLog_$prettydate
-#rm Adium_$prettydate.dmg
-cd ..
-rm .lastadiumbuild
-echo `date +"%Y-%m-%d"` >> .lastadiumbuild
 exit 0
